requirements.txt
pandas
numpy
scikit-learn
imbalanced-learn
xgboost
openai
scipy
azure-ai-ml
azure-identity
